backend/requirements.txt
# Web framework
fastapi
uvicorn[standard]==0.24.0

# Database
sqlalchemy>=2.0.30
alembic==1.12.1
psycopg2-binary  # PostgreSQL driver

# PDF processing
PyMuPDF
pytesseract==0.3.10

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Utilities
pydantic
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
email-validator>=2.0.0
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.25.2
ruff==0.1.6
mypy==1.7.1

# Optional: OCR dependencies (install separately)
# - tesseract (system package)
# - tesseract-language packs (system packages)
